refactoring: Fare를 chain of responsibility로 리펙토링
