+++ src/main/java/nextstep/subway/path/domain/fare/FareChain.java
@@ -1,0 +1,8 @@
+package nextstep.subway.path.domain.fare;
+
+public interface FareChain {
+
+    void setFareChain(FareChain nextChain);
+
+    int calculate(int distance, int fare);
+}
